Add crawl_intchron()
Recursive function for retrieving all records from a given entry point. Called on <https://intchron.org/host> or <https://intchron.org/record> it grabs the whole database in a rough form.

Relies on extra dependencies tools and purrr, but these should be 'free' as tools is bundled with base and purrr is a dependency of tidyr.